project/requirements.txt
django
django-allauth
django-location-field
six
requests